main.py
import pandas as pd
import yaml
from datetime import datetime
import logging
from strategies import your_trading_strategy_function  # Import your trading strategy
from risk_model import calculate_risk_metrics  # Import your risk model
from performance_metrics import calculate_performance_metrics  # Import your performance metrics
from broker_interface import BrokerInterface  # Import your broker interface
from log_generater import setup_logging, log_info

def load_configuration(config_file):
    """
    Load configuration from a YAML file.

    :param config_file: Path to the configuration YAML file
    :return: Dictionary containing configuration settings
    """
    with open(config_file, 'r') as config_yaml:
        config = yaml.safe_load(config_yaml)
    return config

def main():
    # Load configuration settings
    config = load_configuration('config.yaml')

    # Set up logging (customize log file path as needed)
    log_file = f"trading_log_{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.log"
    setup_logging(log_file)

    # Load historical trading data (replace with your data source)
    trading_data = pd.read_csv(config['data_file'])

    # Initialize the broker interface with your API key
    broker = BrokerInterface(api_key=config['api_key'])

    # Execute trading strategy
    log_info("Executing trading strategy...")
    signals = your_trading_strategy_function(trading_data, config)

    # Apply risk model
    log_info("Calculating risk metrics...")
    risk_metrics = calculate_risk_metrics(signals, trading_data, config)

    # Execute trades and manage positions using the broker interface
    log_info("Executing trades...")
    for signal in signals:
        # Implement trade execution logic using the broker interface
        # Example: broker.place_market_order(symbol, quantity, action)

    # Calculate performance metrics
    log_info("Calculating performance metrics...")
    performance_metrics = calculate_performance_metrics(signals, trading_data, config)

    # Log performance metrics
    for key, value in performance_metrics.items():
        log_info(f"{key}: {value:.4f}")

if __name__ == "__main__":
    main()
